finished validating the van der Waals energy for OPLS-AA; moved the NCSRestraint energy to be evaluated at Lambda=0 for DualTopology calculations

+++ modules/potentials/src/main/java/ffx/potential/nonbonded/MaskingInterface.java
@@ -44,7 +44,7 @@ public interface MaskingInterface {
      * @param i The atom whose masking rules should be applied.
      * @since 1.0
      */
-    public void applyMask(final byte mask[], final int i);
+    public void applyMask(final double mask[], final int i);
 
     /**
      * After calling removeMask, all entries in the mask array should be 1.
@@ -53,5 +53,5 @@ public interface MaskingInterface {
      * @param i The atom whose masking rules should be removed.
      * @since 1.0
      */
-    public void removeMask(final byte mask[], final int i);
+    public void removeMask(final double mask[], final int i);
 }
